C7: remove rlearner, relocate outcome-model factory into power_model
Drop the rlearner package entirely, carrying forward only the shared piece
power_model still needs.

- Relocate make_outcome_model (+ its _COMMON params and lazy lightgbm import)
  verbatim from rlearner/nuisance.py into power_model/fitting.py; repoint
  power_model/method.py and inspect_era5_matching_importance.py. Estimator
  construction is byte-identical, so power_model behaviour is unchanged.
- Carry over the factory's direct unit test into test_power_model_fitting.py
  (TestMakeOutcomeModel) so relocating it does not drop its unit coverage.
- Delete the rlearner package, its seven tests, and the rlearner-only
  inspect_prepost_feature_ablation.py. Rename the "rlearner" fixture labels in
  test_plots.py (they were plain strings, no import).
- Drop the now-stale rlearner references in era5_sync.py and method.py docstrings.

Acceptance: poe all-fast green; power_model reads UNCHANGED against both
committed benchmarks (study_toggle_methods_compare max delta 0.06 pp;
study_power_model_compare 0 material moves), i.e. identical to within LightGBM's
same-machine noise floor.
