+++ benchmarking/baselines/power_model/fitting.py
@@ -1,14 +1,21 @@
-"""Time-blocked fold assignment for the power model's baseline holdout fit.
+"""Model-fitting pieces for the power model: fold assignment and the outcome-model factory.
 
 :func:`time_block_folds` — contiguous-block, round-robin fold assignment for time-ordered rows.
 A shuffled split leaks autocorrelation (held-out rows sit minutes from training rows), so its
 residuals are optimistic; contiguous blocks confine the leakage to the block edges, which makes
 the held-out fit-quality diagnostic honest.
+
+:func:`make_outcome_model` — the L2 LightGBM regressor for the counterfactual power ``E[Y|X]``.
 """
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING, Any
+
 import numpy as np
+
+if TYPE_CHECKING:
+    from lightgbm import LGBMRegressor
 
 
 def time_block_folds(n: int, *, n_folds: int = 5, n_blocks: int = 25) -> np.ndarray:
@@ -25,3 +32,32 @@ def time_block_folds(n: int, *, n_folds: int = 5, n_blocks: int = 25) -> np.ndar
         raise ValueError(msg)
     block = np.minimum((np.arange(n) * n_blocks) // max(n, 1), n_blocks - 1)
     return (block % n_folds).astype(int)
+
+
+# Common LightGBM hyperparameters; native NaN handling, seconds to train. Callers (and drivers)
+# override via keyword arguments.
+_COMMON: dict[str, Any] = {
+    "n_estimators": 600,
+    "learning_rate": 0.03,
+    "num_leaves": 63,
+    "min_child_samples": 200,
+    "subsample": 0.8,
+    "colsample_bytree": 0.8,
+    "verbose": -1,
+}
+
+
+def _import_lightgbm() -> Any:  # noqa: ANN401
+    """Import lightgbm lazily with a helpful error if the optional ``ml`` group is missing."""
+    try:
+        import lightgbm  # noqa: PLC0415
+    except ImportError as exc:  # pragma: no cover - exercised only without the optional dep
+        msg = "lightgbm is required for the power model; install the 'ml' optional dependency group."
+        raise ImportError(msg) from exc
+    return lightgbm
+
+
+def make_outcome_model(**overrides: Any) -> LGBMRegressor:  # noqa: ANN401
+    """L2 outcome regressor for the counterfactual power ``E[Y|X]`` (the energy-relevant mean model)."""
+    lgb = _import_lightgbm()
+    return lgb.LGBMRegressor(objective="regression", **{**_COMMON, **overrides})
